+++ packages/gatsby/cache-dir/__tests__/loader.js
@@ -36,7 +36,6 @@ describe(`Production loader`, () => {
 
     const defaultPayload = {
       path: `/mypage/`,
-      webpackCompilationHash: `1234`,
     }
 
     // replace the real XHR object with the mock XHR object before each test
@@ -217,7 +216,6 @@ describe(`Production loader`, () => {
       const prodLoader = new ProdLoader(null, [])
       const payload = {
         path: `/blocked-page/`,
-        webpackCompilationHash: `1234`,
       }
 
       let xhrCount = 0
@@ -261,7 +259,29 @@ describe(`Production loader`, () => {
       }
     }
 
-    beforeEach(() => emitter.emit.mockReset())
+    let originalPathPrefix
+
+    beforeEach(() => {
+      originalPathPrefix = global.__PATH_PREFIX__
+      global.__PATH_PREFIX__ = ``
+      mock.setup()
+      mock.get(`/app-data.json`, (req, res) =>
+        res
+          .status(200)
+          .header(`content-type`, `application/json`)
+          .body(
+            JSON.stringify({
+              webpackCompilationHash: `123`,
+            })
+          )
+      )
+      emitter.emit.mockReset()
+    })
+
+    afterEach(() => {
+      global.__PATH_PREFIX__ = originalPathPrefix
+      mock.teardown()
+    })
 
     it(`should be successful when component can be loaded`, async () => {
       const asyncRequires = createAsyncRequires({
@@ -271,7 +291,6 @@ describe(`Production loader`, () => {
       const pageData = {
         path: `/mypage/`,
         componentChunkName: `chunk`,
-        webpackCompilationHash: `123`,
         result: {
           pageContext: `something something`,
         },
@@ -307,7 +326,6 @@ describe(`Production loader`, () => {
       const pageData = {
         path: `/mypage/`,
         componentChunkName: `chunk`,
-        webpackCompilationHash: `123`,
       }
       prodLoader.loadPageDataJson = jest.fn(() =>
         Promise.resolve({
@@ -335,7 +353,6 @@ describe(`Production loader`, () => {
       const pageData = {
         path: `/mypage/`,
         componentChunkName: `chunk`,
-        webpackCompilationHash: `123`,
       }
       prodLoader.loadPageDataJson = jest.fn(() =>
         Promise.resolve({
@@ -358,7 +375,6 @@ describe(`Production loader`, () => {
       const pageData = {
         path: `/mypage/`,
         componentChunkName: `chunk`,
-        webpackCompilationHash: `123`,
       }
       prodLoader.loadPageDataJson = jest.fn(() =>
         Promise.resolve({